requirements.txt
langchain>=0.0.325
langchain-chroma>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
openai>=1.3.0
chromadb>=0.4.18
pypdf>=3.17.1
beautifulsoup4>=4.12.2
requests>=2.31.0
tqdm>=4.66.1
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.9
pydantic>=2.4.2
schedule>=1.2.0
playwright>=1.40.0
apscheduler>=3.10.4
tiktoken>=0.5.1
aiohttp>=3.8.6
nest_asyncio>=1.5.8
langchain-ollama>=0.0.2
